Spelling error, poorly named variable, bloat.
I fixed grammatical / spelling error in a comment. "Too" not "to."
I changed a poorly named variable. It's not the time remaining. The time remaining is 25 - the variable. It's the age of the last message. 
I removed bloat. A string manipulating ternary condition running every time a user is calling moose too often is more than likely exploitable, and absolutely a bad thing in it's own right.

``` NOTE: I haven't tested these changes, and it appears you don't have tests set up in the repo. You should consider that or run whatever ones you have locally. ```